Redesign Horizontal Apps Slides as widget Renamed it and nested the animation behaviour directly inside the class
