--- Assets/Scripts/UI/HorizontalMenuSlider.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 17f4a08efa4adab48bdf5ace38ed020d
-timeCreated: 1537877085
-licenseType: Free
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
